Source organizing. Second iteration
