--- app/src/main/java/com/github/avanlex/fundamentalsassignments/IMovieListLoader.kt
@@ -1,7 +1,0 @@
-package com.github.avanlex.fundamentalsassignments
-
-import com.github.avanlex.fundamentalsassignments.data.Movie
-
-interface IMovieListLoader {
-    suspend fun getMovies(): List<Movie>
-}
