+++ app/src/main/java/com/github/avanlex/fundamentalsassignments/movieDetails/presentation/ActorsListItemOffsetDecorator.kt
@@ -1,10 +1,10 @@
-package com.github.avanlex.fundamentalsassignments
+package com.github.avanlex.fundamentalsassignments.movieDetails.presentation
 
 import android.graphics.Rect
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 
-class MoviesListItemOffsetDecorator(
+class ActorsListItemOffsetDecorator(
     private val offset: Int
 ): RecyclerView.ItemDecoration() {
 
@@ -14,10 +14,7 @@ class MoviesListItemOffsetDecorator(
         parent: RecyclerView,
         state: RecyclerView.State
     ) {
-        outRect.right = 0
-        outRect.left = offset
-        outRect.top = offset
-        outRect.bottom = offset
+        outRect.right = offset
     }
 
 }
