+++ app/src/main/java/com/github/avanlex/fundamentalsassignments/movieDetails/presentation/FragmentMoviesDetails.kt
@@ -1,4 +1,4 @@
-package com.github.avanlex.fundamentalsassignments
+package com.github.avanlex.fundamentalsassignments.movieDetails.presentation
 
 import android.graphics.ColorFilter
 import android.graphics.ColorMatrix
@@ -14,7 +14,9 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
-import com.github.avanlex.fundamentalsassignments.data.Movie
+import com.github.avanlex.fundamentalsassignments.R
+import com.github.avanlex.fundamentalsassignments.VectorRatingBar
+import com.github.avanlex.fundamentalsassignments.movieList.data.Movie
 
 class   FragmentMoviesDetails : Fragment() {
 
@@ -35,7 +37,7 @@ class   FragmentMoviesDetails : Fragment() {
             .fallback(R.drawable.ic_movie_placeholder)
 
         const val MOVIE_KEY = "MOVIE"
-        fun newInstance(movie: Movie): FragmentMoviesDetails{
+        fun newInstance(movie: Movie): FragmentMoviesDetails {
             val fragment = FragmentMoviesDetails()
             val args = Bundle()
             args.putParcelable(MOVIE_KEY, movie)
