+++ app/src/main/java/com/github/avanlex/fundamentalsassignments/movieList/data/JsonLoad.kt
@@ -1,4 +1,4 @@
-package com.github.avanlex.fundamentalsassignments.data
+package com.github.avanlex.fundamentalsassignments.movieList.data
 
 import android.content.Context
 import kotlinx.coroutines.Dispatchers
