+++ app/src/main/java/com/github/avanlex/fundamentalsassignments/movieList/domain/IMovieListLoader.kt
@@ -1,0 +1,7 @@
+package com.github.avanlex.fundamentalsassignments.movieList.domain
+
+import com.github.avanlex.fundamentalsassignments.movieList.data.Movie
+
+interface IMovieListLoader {
+    suspend fun getMovies(): List<Movie>
+}
