+++ app/src/main/java/com/github/avanlex/fundamentalsassignments/movieList/presentation/MoviesRecyclerViewAdapter.kt
@@ -1,4 +1,4 @@
-package com.github.avanlex.fundamentalsassignments
+package com.github.avanlex.fundamentalsassignments.movieList.presentation
 
 import android.view.LayoutInflater
 import android.view.View
@@ -10,7 +10,10 @@ import androidx.core.graphics.drawable.DrawableCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
-import com.github.avanlex.fundamentalsassignments.data.Movie
+import com.github.avanlex.fundamentalsassignments.R
+import com.github.avanlex.fundamentalsassignments.VectorRatingBar
+import com.github.avanlex.fundamentalsassignments.context
+import com.github.avanlex.fundamentalsassignments.movieList.data.Movie
 import com.google.android.material.imageview.ShapeableImageView
 
 class MoviesRecyclerViewAdapter : RecyclerView.Adapter<MovieViewHolder>() {
@@ -42,7 +45,7 @@ class MoviesRecyclerViewAdapter : RecyclerView.Adapter<MovieViewHolder>() {
     }
 
     override fun onBindViewHolder(holder: MovieViewHolder, position: Int) {
-        holder.onBind(movies[position],
+        holder.bindItem(movies[position],
                 onOpenDetailsClickListener,
                 onAddToFavoriteClickListener
         )
@@ -76,9 +79,9 @@ class MovieViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
             .fallback(R.drawable.ic_movie_placeholder)
     }
 
-    fun onBind(movie: Movie,
-               onOpenDetails: OnItemClickListener?,
-               onAddToFavorite: OnItemAddToFavoriteClickListener?
+    fun bindItem(movie: Movie,
+                 onOpenDetails: OnItemClickListener?,
+                 onAddToFavorite: OnItemAddToFavoriteClickListener?
     ) {
         tagline.text = movie.genres.joinToString { it.name }
         name.text = movie.title
